Move the namespace from retriesapi to api.retries
Also, moved the backoff strategies out of the interface class an into
their own clasess and remove the constants from the interface into its
own class.
