+++ core/retries-api/src/main/java/software/amazon/awssdk/api/retries/AcquireInitialTokenResponse.java
@@ -13,18 +13,18 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.retriesapi;
+package software.amazon.awssdk.api.retries;
 
 import java.time.Duration;
 import software.amazon.awssdk.annotations.SdkPublicApi;
 import software.amazon.awssdk.annotations.ThreadSafe;
 
 /**
- * Response from the {@link RetryStrategy} after calling {@link RetryStrategy#refreshRetryToken(RefreshRetryTokenRequest)}.
+ * Encapsulates the response from the {@link RetryStrategy} to the request to start the attempts to be executed.
  */
 @SdkPublicApi
 @ThreadSafe
-public interface RefreshRetryTokenResponse {
+public interface AcquireInitialTokenResponse {
     /**
      * A {@link RetryToken} acquired by this invocation, used in subsequent {@link RetryStrategy#refreshRetryToken} or
      * {@link RetryStrategy#recordSuccess} calls.
@@ -32,7 +32,7 @@ public interface RefreshRetryTokenResponse {
     RetryToken token();
 
     /**
-     * The amount of time to wait before performing the next attempt.
+     * The amount of time to wait before performing the first attempt.
      */
     Duration delay();
 }
