+++ core/retries-api/src/main/java/software/amazon/awssdk/api/retries/RecordSuccessResponse.java
@@ -13,26 +13,18 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.retriesapi;
+package software.amazon.awssdk.api.retries;
 
-import java.time.Duration;
 import software.amazon.awssdk.annotations.SdkPublicApi;
 import software.amazon.awssdk.annotations.ThreadSafe;
 
 /**
- * Encapsulates the response from the {@link RetryStrategy} to the request to start the attempts to be executed.
+ * Response given to the calling code by the {@link RetryStrategy} after calling
+ * {@link RetryStrategy#recordSuccess(RecordSuccessRequest)}.
  */
 @SdkPublicApi
 @ThreadSafe
-public interface AcquireInitialTokenResponse {
-    /**
-     * A {@link RetryToken} acquired by this invocation, used in subsequent {@link RetryStrategy#refreshRetryToken} or
-     * {@link RetryStrategy#recordSuccess} calls.
-     */
-    RetryToken token();
+public interface RecordSuccessResponse {
 
-    /**
-     * The amount of time to wait before performing the first attempt.
-     */
-    Duration delay();
+    RetryToken token();
 }
