+++ core/retries-api/src/main/java/software/amazon/awssdk/api/retries/RefreshRetryTokenRequest.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.retriesapi;
+package software.amazon.awssdk.api.retries;
 
 import java.time.Duration;
 import java.util.Optional;
