+++ core/retries-api/src/test/java/software/amazon/awssdk/api/retries/ExponentialDelayWithJitterTest.java
@@ -13,20 +13,20 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.retriesapi;
+package software.amazon.awssdk.api.retries;
 
-import static org.assertj.core.api.Assertions.assertThat;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
 
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Random;
 import java.util.function.Function;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 
-@RunWith(Parameterized.class)
 public class ExponentialDelayWithJitterTest {
     static final ComputedNextInt MIN_VALUE_RND = new ComputedNextInt(bound -> 0);
     static final ComputedNextInt MID_VALUE_RND = new ComputedNextInt(bound -> bound / 2);
@@ -34,7 +34,6 @@ public class ExponentialDelayWithJitterTest {
     static final Duration BASE_DELAY = Duration.ofMillis(23);
     static final Duration MAX_DELAY = Duration.ofSeconds(20);
 
-    @Parameterized.Parameters
     public static Collection<TestCase> parameters() {
         return Arrays.asList(
             // --- Using random that returns: bound - 1
@@ -127,12 +126,10 @@ public class ExponentialDelayWithJitterTest {
         );
     }
 
-    @Parameterized.Parameter
-    public TestCase testCase;
-
-    @Test
-    public void testCase() {
-        assertThat(testCase.run()).isEqualTo(testCase.expected());
+    @ParameterizedTest
+    @MethodSource("parameters")
+    public void testCase(TestCase testCase) {
+        assertThat(testCase.run(), equalTo(testCase.expected()));
     }
 
     static class TestCase {
@@ -157,7 +154,7 @@ public class ExponentialDelayWithJitterTest {
 
         Duration run() {
             return
-                new BackoffStrategy.ExponentialDelayWithJitter(() -> random, BASE_DELAY, MAX_DELAY)
+                new ExponentialDelayWithJitter(() -> random, BASE_DELAY, MAX_DELAY)
                     .computeDelay(this.attempt);
         }
