+++ core/retries-api/src/test/java/software/amazon/awssdk/api/retries/FixedDelayWithJitterTest.java
@@ -13,27 +13,25 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.retriesapi;
+package software.amazon.awssdk.api.retries;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
 
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Random;
 import java.util.function.Function;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 
-@RunWith(Parameterized.class)
 public class FixedDelayWithJitterTest {
     static final ComputedNextInt MIN_VALUE_RND = new ComputedNextInt(bound -> 0);
     static final ComputedNextInt MID_VALUE_RND = new ComputedNextInt(bound -> bound / 2);
     static final ComputedNextInt MAX_VALUE_RND = new ComputedNextInt(bound -> bound - 1);
     static final Duration BASE_DELAY = Duration.ofMillis(23);
 
-    @Parameterized.Parameters
     public static Collection<TestCase> parameters() {
         return Arrays.asList(
             // --- Using random that returns: bound - 1
@@ -126,12 +124,11 @@ public class FixedDelayWithJitterTest {
         );
     }
 
-    @Parameterized.Parameter
-    public TestCase testCase;
 
-    @Test
-    public void testCase() {
-        assertThat(testCase.run()).isEqualTo(testCase.expected());
+    @ParameterizedTest
+    @MethodSource("parameters")
+    public void testCase(TestCase testCase) {
+        assertThat(testCase.run(), equalTo(testCase.expected()));
     }
 
     static class TestCase {
@@ -156,7 +153,7 @@ public class FixedDelayWithJitterTest {
 
         Duration run() {
             return
-                new BackoffStrategy.FixedDelayWithJitter(() -> random, BASE_DELAY)
+                new FixedDelayWithJitter(() -> random, BASE_DELAY)
                     .computeDelay(this.attempt);
         }
